Update experrorcalc.py
Added a new input request that asks the user if they wish to import a csv file. If yes, the user selects a csv file from File Explorer, which is then inserted into x_data and y_data, or just x_data if the csv file is a single row. If no, the class passes through the if-else statement and makes use of either the x_data and/or y_data that has been passed in by the user, or the standard initialization arrays. If the user input is neither y/yes or n/no, the class returns that the inputs is unknown then exits the class.

Methods were also updated to not need x_data and y_data passed in, instead making use of data.x_data and data.y_data from the initialization function.